Main Code for Micropython & PC
In recent years, AI has been developing rapidly, especially in the field of deep learning. In this project, we used a deep learning-based image object detection model. After adjusting the model's weights and running batch training, it was able to recognize apples with the accuracy we needed. This allowed us to build a system that classifies apples by combining the model with image capture.

Based on the experience of fruit farmers, we trained the model to classify apples into three grades:

Grade A: high-quality apples

Grade B: medium-quality apples

Grade C: lower-quality apples

We then built a full apple sorting system based on this model. A conveyor belt simulates apples moving into the production line, and a robotic arm handles the apples in an efficient way within the limited space. Finally, apples are sorted into different color-coded packages to show the classification result.